Use direct data copying to combine chunks
Previous approach was causing problems with longer video downloads due
to the size of the command/number of files passed to ffmpeg. This
approach is to sequentially iterate the downloaded chunks and directly
write their content to the output file and now shows a progress bar due
to ease of implementation.